Fix calendar event creation: isCritical/is_critical field mismatch
The form sent camelCase isCritical while the Supabase column is
snake_case is_critical, so every insert was silently rejected by
PostgREST. Also widen calendar_events.duration to NUMERIC(3,1) since
the UI allows half-hour increments but the column was INTEGER.

+++ supabase_schema.sql
@@ -345,7 +345,7 @@ CREATE TABLE calendar_events (
   subject TEXT,
   day INTEGER,
   hour INTEGER,
-  duration INTEGER,
+  duration NUMERIC(3,1),
   type TEXT,
   is_critical BOOLEAN DEFAULT FALSE,
   view_type TEXT CHECK (view_type IN ('individual','group')),
